+++ infra/cdk/app.py
@@ -2,12 +2,11 @@
 """clAWS CDK Application."""
 
 import aws_cdk as cdk
-
+from stacks.gateway_stack import ClawsGatewayStack
+from stacks.guardrails_stack import ClawsGuardrailsStack
+from stacks.policy_stack import ClawsPolicyStack
 from stacks.storage_stack import ClawsStorageStack
 from stacks.tools_stack import ClawsToolsStack
-from stacks.gateway_stack import ClawsGatewayStack
-from stacks.policy_stack import ClawsPolicyStack
-from stacks.guardrails_stack import ClawsGuardrailsStack
 
 app = cdk.App()